Refine component naming
