+++ src/components/icon/index.tsx
@@ -1,0 +1,3 @@
+import Icon from './Icon';
+
+export default Icon;
